BUG: Fix signed/unsigned comparisons in the nifti tools
Three -Wsign-compare warnings, all comparing a signed value against an
unsigned one.  In each case the signed operand is silently converted to
the unsigned type, so a negative value compares as enormous and the
guard it was written to provide does not fire.

  nifti1_tool.c, nifti_tool.c, fill_cmd_string()
      `len < 0 || len >= remain`, where len is int and remain is size_t.
      The `len < 0` test short-circuits first, so the conversion could
      not actually misfire here, but the comparison still relies on that
      ordering to be correct.  Made the conversion explicit, matching the
      idiom already used a few lines above for the first snprintf().

  nifti_tool.c, read_file_text()
      `bytes != len64`, where bytes is size_t from fread() and len64 is
      int64_t.  len64 is validated as > 0 and <= INT_MAX immediately
      above, so the cast is lossless.

No behaviour change; these make the existing intent explicit so the
comparisons no longer depend on the surrounding guards for correctness.

+++ nifti2/nifti_tool.c
@@ -968,7 +968,7 @@ fill_cmd_string(nt_opts * opts, int argc, const char * argv[])
     else
       len = snprintf(cp, remain, " %s", argv[ac]);
 
-    if (len < 0 || len >= remain)
+    if (len < 0 || (size_t)len >= remain)
     {
       fprintf(stderr, "FCS: error parsing command, continuing...\n");
       return 1;
@@ -2370,7 +2370,7 @@ read_file_text(const char * filename, int * length)
   bytes = fread(text, sizeof(char), len64, fp);
   fclose(fp); /* in any case */
 
-  if (bytes != len64)
+  if (bytes != (size_t)len64)
   {
     fprintf(stderr, "** RFT: read only %zu of %" PRId64 " bytes from %s\n", bytes, len64, filename);
     free(text);
